+++ example/lib/network.dart
@@ -1,4 +1,5 @@
 import 'package:http_method/http_method.dart';
+import 'package:component_set/paging/pagination_controller.dart';
 import 'schema.dart';
 import 'myclient.dart';
 part 'network.g.dart';